feat:DRY extractor code
