+++ plugins/json-extractor/package.json
@@ -30,7 +30,7 @@
     "@flatfile/api": "^1.4.9",
     "@flatfile/hooks": "^1.3.0",
     "@flatfile/listener": "^0.3.10",
-    "@flatfile/util-file-buffer": "0.0.2",
+    "@flatfile/util-extractor": "0.1.0",
     "remeda": "^1.14.0"
   },
   "devDependencies": {
