+++ plugins/psv-extractor/src/index.ts
@@ -7,5 +7,5 @@ export const PSVExtractor = (options?: {
   skipEmptyLines?: boolean | 'greedy'
   transform?: (value: any) => Flatfile.CellValueUnion
 }) => {
-  return DelimiterExtractor('|', 'psv', options)
+  return DelimiterExtractor('psv', { delimiter: '|', ...options })
 }
