--- utils/extractor/ref/test.csv
@@ -1,3 +1,0 @@
-name,email,notes
-John Doe,john@doe.com,foobar
-Jane Doe,jane@doe.com,doebar
